requirements.txt
fastapi
uvicorn
gunicorn
redis
faiss-cpu
numpy
pandas
sentence-transformers
python-dotenv
openai
google-generativeai
httpx
python-multipart
pydantic
pydantic-settings
pypdf
python-docx
langchain-text-splitters